TMUP-271: A failed attempt at a redis image.
Due to the lack of a privileged container, ulimit is unavailable and unfortunately is present
in the /etc/init.d/redis6379 service init script.

My attempts to override the template have not been successful due to usage of a chef
resource provider rather than a direct template usage in the recipe.

Usage:
    $ vagrant up redis --provider=docker